Move Athenz mocks to controller-api
